Move `exchanges.js` into `storage` subdir.
