+++ lib/inviteRequest/inviteRequest.js
@@ -1,8 +1,8 @@
 /*!
- * Copyright (c) 2025 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2025-2026 Digital Bazaar, Inc. All rights reserved.
  */
 import * as bedrock from '@bedrock/core';
-import * as exchanges from '../exchanges.js';
+import * as exchanges from '../storage/exchanges.js';
 import {emitExchangeUpdated, evaluateExchangeStep} from '../helpers.js';
 import {logger} from '../logger.js';
