+++ lib/storage/exchanges.js
@@ -1,11 +1,11 @@
 /*!
- * Copyright (c) 2022-2025 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2022-2026 Digital Bazaar, Inc. All rights reserved.
  */
 import * as bedrock from '@bedrock/core';
 import * as database from '@bedrock/mongodb';
-import {parseLocalId, stripStacktrace} from './helpers.js';
+import {parseLocalId, stripStacktrace} from '../helpers.js';
 import assert from 'assert-plus';
-import {logger} from './logger.js';
+import {logger} from '../logger.js';
 import {serializeError} from 'serialize-error';
 
 const {util: {BedrockError}} = bedrock;
@@ -18,11 +18,9 @@ If an exchange is marked as complete, any attempt to mark it complete again
 will result in an action, as specified in the exchange record, being taken
 such as auto-revocation or notification.
 
-Each pending exchange may include optionally encrypted VCs for pickup and / or
-VC templates and required variables (which may come from other VCs) that must
-be provided to populate those templates. If any templates are provided, then
-a capability to issue the VC must also be provided. If any VCs are to be
-provided during the exchange a capability to verify them must be provided. */
+Each pending exchange is an instance of a workflow. A workflow may have
+one or more steps, each that might issue, verify, or deliver VCs. Capabilities
+must be provided to issue or verify VCs. */
 
 const COLLECTION_NAME = 'vc-exchange';
