+++ lib/vcapi.js
@@ -1,8 +1,8 @@
 /*!
- * Copyright (c) 2018-2025 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2018-2026 Digital Bazaar, Inc. All rights reserved.
  */
 import * as bedrock from '@bedrock/core';
-import * as exchanges from './exchanges.js';
+import * as exchanges from './storage/exchanges.js';
 import {createChallenge as _createChallenge, verify} from './verify.js';
 import {
   buildPresentationFromResults,
